% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucrdtw.R
\name{summary.ucrdtw}
\alias{summary.ucrdtw}
\title{Summarize subsequence search}
\usage{
\method{summary}{ucrdtw}(object, ...)
}
\arguments{
\item{object}{an object of class ucrdtw}

\item{...}{further arguments passed to or from methods}
}
\value{
An unlisted version of the object is returned.
}
\description{
Summary method for class ucrdtw
}
