% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucrdtw.R
\name{summary.ucred}
\alias{summary.ucred}
\title{Summarize subsequence search}
\usage{
\method{summary}{ucred}(object, ...)
}
\arguments{
\item{object}{an object of class ucred}

\item{...}{further arguments passed to or from methods}
}
\value{
An unlisted version of the object is returned.
}
\description{
Summary method for class ucred
}

