% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ucrdtw_ff}
\alias{ucrdtw_ff}
\title{UCR DTW Algorithm file-file method}
\usage{
ucrdtw_ff(data, query, qlength, dtwwindow)
}
\arguments{
\item{data}{char path to data file}

\item{query}{char path to query file}

\item{qlength}{int length of query (n data points)}

\item{dtwwindow}{double warping window size (as a proportion of query length)}
}
\description{
This implementation is very close to the UCR Suite command line utility, in that it takes files as inputs for both query and data
}

