% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ucrdtw_fv}
\alias{ucrdtw_fv}
\title{UCR DTW Algorithm file-vector method}
\usage{
ucrdtw_fv(data, query, qlength, dtwwindow)
}
\arguments{
\item{data}{char path to data file}

\item{query}{numeric vector containing the query}

\item{qlength}{int length of query (n data points)}

\item{dtwwindow}{double warping window size (as a proportion of query length)}
}
\description{
This implementation of the UCR Suite command line utility, takes a data file as input and an R numeric vector for the query
}

