% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ucrdtw_vv}
\alias{ucrdtw_vv}
\title{UCR DTW Algorithm vector-vector method}
\usage{
ucrdtw_vv(data, query, qlength, dtwwindow)
}
\arguments{
\item{query}{numeric vector containing the query}

\item{qlength}{int length of query (n data points)}

\item{dtwwindow}{double warping window size (as a proportion of query length)}

\item{datav}{numeric vector containing data}
}
\description{
This implementation of the UCR Suite command line utility, takes an R numeric vector as data input and an R numeric vector for the query
}

