% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucrdtw.R
\name{ucred_vm}
\alias{ucred_vm}
\title{UCR ED Algorithm vector-matrix method}
\usage{
ucred_vm(data, query, qlength, skip = TRUE, byrow = FALSE)
}
\arguments{
\item{data}{numeric matrix containing data}

\item{query}{numeric vector containing the query}

\item{qlength}{int length of query (n data points)}

\item{skip}{bool defaults to TRUE. If TRUE bound calculations and if necessary, distance calculations, are only performed on non-overlapping segments of the data (i.e. multiples of \code{qlength}). This is useful if \code{data} is a set of multiple reference time series, each of length \code{qlength}. The location returned when skipping is the index of the subsequence.}

\item{byrow}{logical If TRUE rows represent time-series, columns time-points}
}
\value{
a ucrdtw object, note that the field \code{location} indicates the row or column number of the reference set
}
\description{
This implementation of the UCR Suite command line utility, takes an R numeric matrix as data input and an R numeric vector for the query.
The default behaviour differs from the other metjods, in that it does not perform a sliding window search for a match.
Instead it is designed to find a best match for a query in a reference set of time-series of the same length as the query.
This is useful, for example, when comparing a time-series of undetermined class to a labelled reference set of classified time-series.
}

