% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ucred_vv}
\alias{ucred_vv}
\title{UCR ED Algorithm vector-vector method}
\usage{
ucred_vv(data, query, qlength)
}
\arguments{
\item{data}{numeric vector containing data}

\item{query}{numeric vector containing query}

\item{qlength}{int length of query (n data points)}
}
\description{
This implementation of the UCR Suite Euclidean Distance command line utility takes an R numeric vector as data input and an R numeric vector for the query.
}

