% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ucred_ff}
\alias{ucred_ff}
\title{UCR ED Algorithm file-file method}
\usage{
ucred_ff(data, query, qlength)
}
\arguments{
\item{data}{char path to data file}

\item{query}{char path to query file}

\item{qlength}{int length of query (n data points)}
}
\description{
This implementation is very close to the UCR Suite command line utility, in that it takes files as inputs for both query and data
}

